Add associate-public-ip-address key to instance resource
The option is supported by RunInstances, CreateSpotInstanceRequest and
CreateLaunchConfiguration. For the latter twp API calls, it's a
top-level instance property, while RunInstances sees the option
explicitly associated with network device zero -- effectively making it
an instance property, again. So, top-level it is here as well.

This commit adds only validation of the key in the instance definition.
It is not propagated to the API calls themselves.